azure-functions/requirements.txt
azure-functions
pandas
pyarrow
numpy
